+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/profiling/pprof/IPprofTaskLogQueryDAO.java
@@ -25,7 +25,11 @@ import org.apache.skywalking.oap.server.core.storage.DAO;
 
 public interface IPprofTaskLogQueryDAO extends DAO {
     /**
-     * search all task log list in appoint task id
+     * Search task logs by the given task id.
+     *
+     * @param taskId the task id to filter by, must not be null or blank
+     * @return the task logs associated with the given task id
+     * @throws IOException if the query fails
      */
-    List<PprofTaskLog> getTaskLogList() throws IOException;
+    List<PprofTaskLog> getTaskLogList(String taskId) throws IOException;
 }